+++ src/pagy.ts
@@ -8,36 +8,17 @@ interface Tags {
     readonly gap:    string
     readonly after:  string
 }
-interface Sequels      { readonly [width:string]: (string|number|"gap")[] }
+interface Sequels      { readonly [width:string]: (string|number)[] }
 interface LabelSequels { readonly [width:string]: string[] }
 interface NavElement extends Element { pagyRender(): void }
 
-const Pagy = {
-    version: "5.8.1",
-
-    // Scan for elements with a "data-pagy" attribute and call their init functions with the decoded args
-    init(arg?:Element|never) {
-        const target   = arg instanceof Element ? arg : document;
-        const elements = target.querySelectorAll("[data-pagy]");
-        for (const el of elements) {
-            try {
-                const uint8array         = Uint8Array.from(atob(el.getAttribute("data-pagy") as string), c => c.charCodeAt(0));
-                const [keyword, ...args] = JSON.parse(new TextDecoder().decode(uint8array)); // base64-utf8 JSON -> Array
-                if (keyword === "nav") {
-                    Pagy.initNav(el as NavElement, args as NavArgs);
-                } else if (keyword === "combo") {
-                    Pagy.initCombo(el, args as ComboArgs);
-                } else if (keyword === "selector") {
-                    Pagy.initSelector(el, args as SelectorArgs);
-                } else {
-                    console.warn("Skipped Pagy.init() for: %o\nUnknown keyword '%s'", el, keyword);
-                }
-            } catch (err) { console.warn("Skipped Pagy.init() for: %o\n%s", el, err) }
-        }
-    },
+const Pagy = (() => {
+    // The observer instance for responsive navs
+    const rjsObserver = new ResizeObserver(entries =>
+        entries.forEach(e => e.target.querySelectorAll<NavElement>(".pagy-rjs").forEach(el => el.pagyRender())));
 
     // Init the *_nav_js helpers
-    initNav(el:NavElement, [tags, sequels, labelSequels, trimParam]:NavArgs) {
+    const initNav = (el:NavElement, [tags, sequels, labelSequels, trimParam]:NavArgs) => {
         const container = el.parentElement ?? el;
         const widths    = Object.keys(sequels).map(w => parseInt(w)).sort((a, b) => b - a);
         let lastWidth   = -1;
@@ -53,7 +34,7 @@ const Pagy = {
                 const item  = series[i];
                 const label = labels[i];
                 if (typeof trimParam === "string" && item === 1) {
-                    html += Pagy.trim(fillIn(tags.link, item.toString(), label), trimParam);
+                    html += trim(fillIn(tags.link, item.toString(), label), trimParam);
                 } else if (typeof item === "number") {
                     html += fillIn(tags.link, item.toString(), label);
                 } else if (item === "gap") {
@@ -67,30 +48,24 @@ const Pagy = {
             el.insertAdjacentHTML("afterbegin", html);
             lastWidth = width;
         })();
-        if (el.classList.contains("pagy-rjs")) { Pagy.rjsObserver.observe(container) }
-    },
-
-    // The observer instance for responsive navs
-    rjsObserver: new ResizeObserver(entries => {
-        entries.forEach(e => e.target.querySelectorAll<NavElement>(".pagy-rjs").forEach(el => el.pagyRender()));
-    }),
+        if (el.classList.contains("pagy-rjs")) { rjsObserver.observe(container) }
+    };
 
     // Init the *_combo_nav_js helpers
-    initCombo(el:Element, [link, trimParam]:ComboArgs) {
-        Pagy.initInput(el, inputValue => [inputValue, link.replace(/__pagy_page__/, inputValue)], trimParam);
-    },
+    const initCombo = (el:Element, [link, trimParam]:ComboArgs) =>
+        initInput(el, inputValue => [inputValue, link.replace(/__pagy_page__/, inputValue)], trimParam);
 
-    // Init the items_selector_js helper
-    initSelector(el:Element, [from, link, trimParam]:SelectorArgs) {
-        Pagy.initInput(el, inputValue => {
+        // Init the items_selector_js helper
+    const initSelector = (el:Element, [from, link, trimParam]:SelectorArgs) => {
+        initInput(el, inputValue => {
             const page = Math.max(Math.ceil(from / parseInt(inputValue)), 1).toString();
             const html = link.replace(/__pagy_page__/, page).replace(/__pagy_items__/, inputValue);
             return [page, html];
         }, trimParam);
-    },
+    };
 
     // Init the input element
-    initInput(el:Element, getVars:(v:string)=>[string, string], trimParam?:string) {
+    const initInput = (el:Element, getVars:(v:string)=>[string, string], trimParam?:string) => {
         const input   = el.querySelector("input") as HTMLInputElement;
         const initial = input.value;
         const action  = function() {
@@ -102,15 +77,41 @@ const Pagy = {
                 return;
             }
             let [page, html] = getVars(input.value);   // eslint-disable-line prefer-const
-            if (typeof trimParam === "string" && page === "1") { html = Pagy.trim(html, trimParam) }
+            if (typeof trimParam === "string" && page === "1") { html = trim(html, trimParam) }
             el.insertAdjacentHTML("afterbegin", html);
             (el.querySelector("a") as HTMLAnchorElement).click();
         };
         ["change", "focus"].forEach(e => input.addEventListener(e, input.select));        // auto-select
         input.addEventListener("focusout", action);                                       // trigger action
         input.addEventListener("keypress", e => { if (e.key === "Enter") { action() } }); // trigger action
-    },
+    };
 
     // Trim the ${page-param}=1 params in links
-    trim: (link:string, param:string) => link.replace(new RegExp(`[?&]${param}=1\\b(?!&)|\\b${param}=1&`), "")
-};
+    const trim = (link:string, param:string) => link.replace(new RegExp(`[?&]${param}=1\\b(?!&)|\\b${param}=1&`), "");
+
+    // Public interface
+    return {
+        version: "5.9.0",
+
+        // Scan for elements with a "data-pagy" attribute and call their init functions with the decoded args
+        init(arg?:Element|never) {
+            const target   = arg instanceof Element ? arg : document;
+            const elements = target.querySelectorAll("[data-pagy]");
+            for (const el of elements) {
+                try {
+                    const uint8array         = Uint8Array.from(atob(el.getAttribute("data-pagy") as string), c => c.charCodeAt(0));
+                    const [keyword, ...args] = JSON.parse((new TextDecoder()).decode(uint8array)); // base64-utf8 -> JSON -> Array
+                    if (keyword === "nav") {
+                        initNav(el as NavElement, args as NavArgs);
+                    } else if (keyword === "combo") {
+                        initCombo(el, args as ComboArgs);
+                    } else if (keyword === "selector") {
+                        initSelector(el, args as SelectorArgs);
+                    } else {
+                        console.warn("Skipped Pagy.init() for: %o\nUnknown keyword '%s'", el, keyword);
+                    }
+                } catch (err) { console.warn("Skipped Pagy.init() for: %o\n%s", el, err) }
+            }
+        }
+    };
+})();